assignment-6/requirements.txt
Flask==1.1.2
google-cloud-datastore==1.15.3
json2html==1.3.0
